Move back watcher to the model.
