+++ src/mainwindow.cpp
@@ -129,6 +129,9 @@ MainWindow::MainWindow()
             });
         }
     });
+
+    connect(m_data_model, &TasksModel::globalUrgencyChanged, m_tray_icon,
+            &SystemTrayIcon::updateStatusIcon);
 }
 
 MainWindow::~MainWindow() { m_task_provider.reset(); }
